rpc protocol

+++ buffer.go
@@ -2,35 +2,35 @@ package eio
 
 type MessageByteBuffer struct {
 	b   []byte
-	len int64
+	len int
 }
 
-func (messageByteBuffer *MessageByteBuffer) Len() int64 {
+func (messageByteBuffer *MessageByteBuffer) Len() int {
 	return messageByteBuffer.len
 }
 
-func (messageByteBuffer *MessageByteBuffer) Discard(l int64) {
+func (messageByteBuffer *MessageByteBuffer) Discard(l int) {
 	messageByteBuffer.b = messageByteBuffer.b[l:]
 	messageByteBuffer.len -= l
 }
 
 func (messageByteBuffer *MessageByteBuffer) Append(b []byte) {
 	messageByteBuffer.b = append(messageByteBuffer.b, b...)
-	messageByteBuffer.len += int64(len(b))
+	messageByteBuffer.len += len(b)
 }
 
 func (messageByteBuffer *MessageByteBuffer) AppendByte(b byte) {
 	messageByteBuffer.Append([]byte{b})
 }
 
-func (messageByteBuffer *MessageByteBuffer) Peek(start int64, end int64) *MessageByteBuffer {
+func (messageByteBuffer *MessageByteBuffer) Peek(start int, end int) *MessageByteBuffer {
 	return &MessageByteBuffer{
 		b:   messageByteBuffer.b[start:end],
 		len: end - start,
 	}
 }
 
-func (messageByteBuffer *MessageByteBuffer) Int64Value(start int64, end int64) int64 {
+func (messageByteBuffer *MessageByteBuffer) Int64Value(start int, end int) int64 {
 	b := messageByteBuffer.b[start:end]
 	return BytesToInt64(b)
 }
